recognition/arcface_torch/requirement.txt
tensorboard
easydict
mxnet
onnx
sklearn
opencv-python
open-clip-torch
transformers
numpy==1.23.1
scikit-learn